Refactor: Integrate VectorBT into BayesianTPEOptimizer
This commit integrates VectorBT's advanced optimization capabilities directly into the BayesianTPEOptimizer. It enhances grid generation and evaluation by leveraging VectorBT's vectorized operations and unified vectorization manager. The separate `EnhancedGridTPEOptimizer` and `VectorBTGridSearch` classes have been removed and their functionalities merged into the main optimizer. This streamlines the codebase and improves performance by utilizing VectorBT more effectively.
